# Avoid a contiguous call in the quantized phi 3 model.