This call did not seem to be necessary and removing it result in some large speedups. When generating ~500 tokens.
- Cuda (4080 super):  161 tok/s -> 180 tok/s.
- Cpu (7950X): 16.4 tok/s -> 16.7 tok/s.
- Metal (MBP M2 16GB 14,9): 35.3 tok/s -> 44.1 tok/s.
- Cpu (MBP M2 16GB 14,9): 18.3 tok/s -> 19.4 tok/s.